Add files via upload
A Smart Water System is designed to improve the management and conservation of water resources through the use of advanced technology and data analytics. The objectives of implementing such a system can vary depending on the specific goals and needs of the community or organization, but here are some common objectives.